src/book_editor/core/enforcement_strategy.py
"""Enforcement strategy implementation for cursor rules."""

import json
import logging
from typing import Dict, Any, Optional
from dataclasses import dataclass
from pathlib import Path

logging.basicConfig(level=logging.INFO)
logger = logging.getLogger(__name__)


@dataclass
class EnforcementMetrics:
    """Metrics for enforcement validation."""
    token_reduction: float = 0.0
    response_time: float = 0.0
    cost_reduction: float = 0.0
    resource_efficiency: float = 0.0


class EnforcementStrategy:
    """Implements the cursor rules enforcement strategy."""
    
    def __init__(self, config_path: Optional[str] = None):
        self.config_path = config_path or str(
            Path(__file__).parent.parent / "config" / "settings.json"
        )
        self.config = self._load_config()
        self.metrics = EnforcementMetrics()
        
    def _load_config(self) -> Dict[str, Any]:
        """Load configuration from file."""
        try:
            with open(self.config_path, 'r') as f:
                return json.load(f)
        except Exception as e:
            logger.error(f"Failed to load config: {e}")
            return {}

    def validate_compliance(self) -> bool:
        """Validate compliance with enforcement rules."""
        try:
            if not self.config or 'metrics_threshold' not in self.config:
                return False
                
            thresholds = self.config['metrics_threshold']
            return all([
                self.metrics.token_reduction >= thresholds.get('token_reduction', 0.35),
                self.metrics.response_time >= thresholds.get('response_time', 0.25),
                self.metrics.cost_reduction >= thresholds.get('cost_reduction', 0.40),
                self.metrics.resource_efficiency >= thresholds.get('resource_efficiency', 0.35)
            ])
        except Exception as e:
            logger.error(f"Compliance validation failed: {e}")
            return False

    def enforce_rules(self) -> bool:
        """Enforce cursor rules based on configuration."""
        try:
            # If config is empty, enforcement fails
            if not self.config:
                logger.warning("Cannot enforce rules with empty configuration")
                return False

            # Always validate compliance
            is_compliant = self.validate_compliance()
            
            # If strict enforcement is enabled and validation fails, return False
            if self.config.get('strict_enforcement', True) and not is_compliant:
                logger.warning("Compliance validation failed")
                return False

            # Performance monitoring
            self._monitor_performance()
            
            # Cost optimization
            self._optimize_costs()
            
            # Return validation result
            return is_compliant
        except Exception as e:
            logger.error(f"Rule enforcement failed: {e}")
            return False

    def _monitor_performance(self):
        """Monitor system performance metrics."""
        metrics_config = self.config.get('metrics_collection', {})
        
        if metrics_config.get('token_usage') == 'per_request':
            self._track_token_usage()
        
        if metrics_config.get('performance_metrics') == 'detailed':
            self._track_performance_metrics()

    def _track_token_usage(self):
        """Track token usage metrics."""
        # Implementation for token tracking
        pass

    def _track_performance_metrics(self):
        """Track detailed performance metrics."""
        # Implementation for performance tracking
        pass

    def _optimize_costs(self):
        """Implement cost optimization strategies."""
        cost_config = self.config.get('cost_control', {})
        
        if cost_config.get('token_efficiency') == 'maximum':
            self._optimize_token_usage()
            
        if cost_config.get('cache_optimization') == 'aggressive':
            self._optimize_cache()

    def _optimize_token_usage(self):
        """Optimize token usage."""
        # Implementation for token optimization
        pass

    def _optimize_cache(self):
        """Optimize cache usage."""
        # Implementation for cache optimization
        pass

    def update_metrics(self, metrics: EnforcementMetrics):
        """Update enforcement metrics."""
        self.metrics = metrics
        self._log_metrics()

    def _log_metrics(self):
        """Log current metrics."""
        logger.info(f"Current Metrics: {self.metrics}")

    def get_status(self) -> Dict[str, Any]:
        """Get current enforcement status."""
        return {
            "compliant": self.validate_compliance(),
            "metrics": {
                "token_reduction": self.metrics.token_reduction,
                "response_time": self.metrics.response_time,
                "cost_reduction": self.metrics.cost_reduction,
                "resource_efficiency": self.metrics.resource_efficiency
            }
        }
